fix(host-health): simplify host health main.go
